Remove guided tours and load it in the template
