--- plugins/system/guidedtours/guidedtours.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<extension version="4.0" type="plugin" group="system" method="upgrade">
-    <name>plg_system_guidedtours</name>
-    <author>Joomla! Project</author>
-    <creationDate>2023-02</creationDate>
-    <copyright>(C) 2023 Open Source Matters, Inc.</copyright>
-    <license>GNU General Public License version 2 or later; see LICENSE.txt</license>
-    <authorEmail>admin@joomla.org</authorEmail>
-    <authorUrl>www.joomla.org</authorUrl>
-    <version>4.3.0</version>
-    <description>PLG_SYSTEM_GUIDEDTOURS_DESCRIPTION</description>
-	<namespace path="src">Joomla\Plugin\System\GuidedTours</namespace>
-    <files>
-		<folder>services</folder>
-		<folder>src</folder>
-    </files>
-    <languages>
-        <language tag="en-GB">language/en-GB/plg_system_guidedtours.ini</language>
-        <language tag="en-GB">language/en-GB/plg_system_guidedtours.sys.ini</language>
-    </languages>
-</extension>
